% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unigram_freq-data.R
\docType{data}
\name{unigram_freq}
\alias{unigram_freq}
\title{English Word Frequency}
\format{
An object of class \code{data.frame} with 333,333 rows and 2 columns. The first column contains words; the second contains their counts.
}
\source{
\href{https://www.kaggle.com/datasets/rtatman/english-word-frequency}{Kaggle}
}
\usage{
data(unigram_freq)
}
\description{
⅓ Million Most Frequent English Words on the Web.
}
\details{
This dataset contains the counts of the 333,333 most commonly-used single words on the English language web, as derived from the Google Web Trillion Word Corpus.
}
\references{
Rachael Tatman (2019)
}
\keyword{datasets}
